Update the node tests structure
